Implemented members page in Members.svelte, Started implementing MemberViewer

+++ backend/src/routes/membersController.ts
@@ -25,6 +25,11 @@ membersRouter.get("/all", async (_, resp) => {
 });
 
 membersRouter.post("/add", async (req, resp) => {
+    // TODO: Send email with temporary password to given email
+
+    let memberId: string | null = null;
+    let memberRev: string | null = null;
+
     try {
         const { name, surname, email, phone, address, voice, status, role, birthdate, joined } = req.body;
 
@@ -32,8 +37,10 @@ membersRouter.post("/add", async (req, resp) => {
         if (!validInputs(name, surname, email, phone, address, voice, status, role, birthdate, joined))
             return resp.status(400).json({ errorMessage: "InvalidInputs" });
 
+        let tempPassword = await generateTempPassword();
+
         // Create new member
-        await dbService.create("members", {
+        const member = await dbService.create("members", {
             name: name,
             surname: surname,
             email: email,
@@ -46,23 +53,32 @@ membersRouter.post("/add", async (req, resp) => {
             joined: joined
         });
 
+        memberId = member.id;
+        memberRev = member.rev;
+
         // Add actual user account
         await dbService.create("users", {
             email: email,
             name: `${name} ${surname}`,
-            password: await hash(await generateTempPassword(), 12),
+            password: await hash(tempPassword, 12),
             phone: phone,
             address: address,
             changePassword: true,
             firstLogin: true,
             userId: uuidv4(),
-            role: role
+            role: role,
+            memberId: memberId
         });
-
-        // TODO: Send email with temporary password to given email
 
         return resp.status(200).json({ ok: true });
     } catch (err: any) {
+        if (memberId && memberRev) {
+            try {
+                await dbService.delete("members", memberId, memberRev);
+            } catch (cleanupErr) {
+                logger.error({ cleanupErr }, "Rollback failed");
+            }
+        }
         logger.error({ err }, "members/add route errorMessage: ");
         return resp.status(500).json({ errorMessage: "InternalServerError" });
     }
@@ -72,11 +88,18 @@ membersRouter.post("/delete", async (req, resp) => {
     try {
         const { id } = req.body;
 
+        if (!id) return resp.status(400).json({ errorMessage: "InvalidInputs" });
+
         const member = await dbService.read("members", id);
+        const users = await dbService.find("users", { selector: { memberId: id }, limit: 1 });
+
+        const user = users[0];
 
         if (!member) return resp.status(404).json({ errorMessage: "MemberNotFound" });
+        // if (!user) return resp.status(404).json({ errorMessage: "UserNotFound" });
 
         await dbService.delete("members", member._id, member._rev);
+        // await dbService.delete("users", user._id, user._rev);
 
         return resp.status(200).json({ ok: true });
     } catch (err: any) {
@@ -86,6 +109,7 @@ membersRouter.post("/delete", async (req, resp) => {
 });
 
 membersRouter.post("/update", async (req, resp) => {
+    // TODO: Update user data
     try {
         const { updatedMember } = req.body;
         if (!updatedMember?.id) {
@@ -120,6 +144,26 @@ membersRouter.post("/update", async (req, resp) => {
     }
 });
 
+membersRouter.post("/update/status", async (req, resp) => {
+    try {
+        const { id } = req.body;
+        if (!id) return resp.status(400).json({ errorMessage: "InvalidInputs" });
+
+        const member = await dbService.read("members", id);
+        if (!member) return resp.status(404).json({ errorMessage: "MemberNotFound" });
+
+        if (member.status === "active") {
+            await dbService.update("members", { ...member, status: "inactive" });
+        } else {
+            await dbService.update("members", { ...member, status: "active" });
+        }
+
+        return resp.status(200).json({ ok: true });
+    } catch (err: any) {
+        logger.error({ err }, "members/update/status route errorMessage");
+        return resp.status(500).json({ errorMessage: "InternalServerError" });
+    }
+});
 
 function validInputs(name: string, surname: string, email: string, phone: string, address: string, voice: string, status: string, role: string, birthdate: string, joined: number) {
     if (!name || !surname || !email || !phone || !address || !voice || !status || !role || !birthdate || !joined || name.trim() === "" || surname.trim() === "") {
